Rename: 자격증 후기 dto 클래스 request와 response패키지로 분리
Related to: #47
