+++ src/main/java/pulleydoreurae/careerquestbackend/certification/domain/dto/request/InterestedCertificationRequest.java
@@ -1,4 +1,4 @@
-package pulleydoreurae.careerquestbackend.certification.domain.dto;
+package pulleydoreurae.careerquestbackend.certification.domain.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
