+++ src/main/java/pulleydoreurae/careerquestbackend/certification/domain/dto/response/CertificationDateResponse.java
@@ -1,4 +1,4 @@
-package pulleydoreurae.careerquestbackend.certification.domain.dto;
+package pulleydoreurae.careerquestbackend.certification.domain.dto.response;
 
 import java.util.ArrayList;
 import java.util.List;
